+++ packages/core/src/electron-main/electron-main-window-service-impl.ts
@@ -18,8 +18,7 @@ import { shell } from '../../shared/electron';
 import { injectable, inject } from 'inversify';
 import { ElectronMainWindowService } from '../electron-common/electron-main-window-service';
 import { ElectronMainApplication } from './electron-main-application';
-// eslint-disable-next-line @theia/runtime-import-check
-import { NewWindowOptions } from '../browser/window/window-service';
+import { NewWindowOptions } from '../common/window';
 
 @injectable()
 export class ElectronMainWindowServiceImpl implements ElectronMainWindowService {