+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - [task] `TaskDefinition.properties.required` is now optional to align with the specification [#10015](https://github.com/eclipse-theia/theia/pull/10015)
 - [core] `setTopPanelVisibily` renamed to `setTopPanelVisibility` [#10020](https://github.com/eclipse-theia/theia/pull/10020)
 - [application-manager] break `rebuild` API: second argument is now an optional object instead of an optional array.
+- [core] moved `DEFAULT_WINDOW_HASH` to `common/window.ts` []()
+- [core] moved `NewWindowOptions` to `common/window.ts` []()
 
 ## v1.17.2 - 9/1/2021